requirements.txt
Django==1.7.4
Pillow==2.7.0
Unidecode==0.04.17
dj-database-url==0.3.0
dj-static==0.0.6
django-filter==0.9.2
djangorestframework==3.0.5
static3==0.5.1
gunicorn
wsgiref